pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "project-spark"
version = "0.1.0"
description = "Spark No. 9 - Client Onboarding System for Snowflake"
authors = [
    { name = "Spark No. 9 Team" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "streamlit>=1.28.0",
    "pyyaml>=6.0.0",
    "snowflake>=1.0.2",
    "snowflake-snowpark-python>=1.30.0",
    "sqlalchemy>=2.0.43",
    "snowflake-sqlalchemy>=1.7.7",
    "snowflake-cli-labs>=0.2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
spark-create-schema = "project_spark.main:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["snowflake.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ""

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
